Moved code into nomenclature wise directories
