+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "portfolio-frontend",
-  "version": "0.2.1",
+  "version": "0.3.1",
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
