+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './sections'
+
+export * from './ui'
