+++ src/components/sections/index.ts
@@ -1,0 +1,1 @@
+export * from './Service/Service'
