+++ src/components/ui/index.ts
@@ -1,0 +1,5 @@
+export * from './SectionHeading/SectionHeading'
+
+export * from './SectionSubHeading/SectionSubHeading'
+
+export * from './SectionTextContent/SectionTextContent'
